requirements.txt
ipylab==0.7.1
ipython-genutils==0.2.0
ipywidgets==8.1.0
jupyterlab-widgets==3.0.8
matplotlib==3.7.1
numpy==1.25.0
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
statsmodels==0.14.0
varname==0.11.2
stargazer
altair==5.1.2
traitlets==5.7.1
jup